delete dead code and move non-generic utilities from shared shared package
