+++ pkg/controller/OWNERS
@@ -7,8 +7,5 @@ reviewers:
   - jim-minter
   - csrwng
 approvers:
-  - smarterclayton
-  - deads2k
   - kargakis
   - soltysh
-  - bparees
