+++ pkg/deploy/controller/deploymentconfig/controller_ref_manager.go
@@ -1,4 +1,4 @@
-package controller
+package deploymentconfig
 
 import (
 	"fmt"
